Split test class into multiple ones, on a per-method basis
